Make exceptions public top-level classes
