+++ jgrapht-core/src/main/java/org/jgrapht/graph/InvalidGraphWalkException.java
@@ -1,0 +1,34 @@
+/*
+ * (C) Copyright 2017-2021, by Joris Kinable and Contributors.
+ *
+ * JGraphT : a free Java graph-theory library
+ *
+ * See the CONTRIBUTORS.md file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the
+ * GNU Lesser General Public License v2.1 or later
+ * which is available at
+ * http://www.gnu.org/licenses/old-licenses/lgpl-2.1-standalone.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR LGPL-2.1-or-later
+ */
+package org.jgrapht.graph;
+
+/**
+ * Exception thrown in the event that the path is invalid.
+ */
+public class InvalidGraphWalkException
+    extends
+    RuntimeException
+{
+    private static final long serialVersionUID = 3811666107707436479L;
+
+    public InvalidGraphWalkException(String message)
+    {
+        super(message);
+    }
+
+}
