+++ jgrapht-core/src/main/java/org/jgrapht/util/SupplierUtil.java
@@ -252,16 +252,4 @@ public class SupplierUtil
             return new SerializedForm<>(constructor.getDeclaringClass());
         }
     }
-
-    static class SupplierException
-        extends
-        IllegalArgumentException
-    {
-        private static final long serialVersionUID = -8192314371524515620L;
-
-        public SupplierException(String message, Throwable cause)
-        {
-            super(message, cause);
-        }
-    }
 }
